requirements.txt
PyYAML
rich
pdf2image
Pillow
opencv-python
numpy
scikit-image
pytesseract
layoutparser
# Recommend installing torch separately first, but include for completeness
# torch==2.2.1
# torchvision==0.17.1
# For layoutparser[detectron2], Detectron2 must be installed manually.
# Post-processing & Output
symspellpy
pandas
markdown-it-py
PyMuPDF
dask[complete]
